feat: Add Bash script for compiling C source files
This adds a Bash script that automates the compilation of C source files.
The script takes a filename as an argument and checks if the provided file has a '.c' extension.
If so, it uses the 'gcc' compiler to generate an executable with the same name (without the '.c' extension).
If the file is not a C source file, the script exits with an error message.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# Date:      Thu Aug 31 14:11:29 2023 +0100
#
# On branch main
# Your branch is up to date with 'origin/main'.
#
# Changes to be committed:
#	new file:   c_compiler.sh
#
# Untracked files:
#	script.sh
#